выполнил 11-12 задания,  немного исправил предыдущие (int заменил на double)

+++ src/main/java/task_6/Rectangle.java
@@ -1,14 +1,14 @@
 package task_6;
 
 public class Rectangle {
-    int width;
-    int height;
+    double width;
+    double height;
 
-    public int area(){
+    public double area(){
         return width*height;
     }
 
-    public int perimeter() {
+    public double perimeter() {
         return 2*(width+height);
     }
